refactor(shell): extract ExportController from ReaderActivity
Move PDF annotation export (ADR-INKREAD-0005) — the editable/flatten chooser,
the all-files-access gate, the engine-thread write, and the synced-root search
(findOriginalParent + EXPORT_DIR_NAME/SYNCED_DIRS) — into a focused
ExportController behind a narrow Host seam (activity/docHandle/currentDocPath/
engineExecute). The toolbar now calls export.showExportDialog().

Behavior + threading preserved. ReaderActivity drops ~90 more lines (2617).
Third controller extracted (Search, Dict, Export). Kotlin compiles.
